+++ dropback/run.sh
@@ -60,8 +60,21 @@ get_dropbox_file_path() {
     fi
 }
 
+# create/update HASS entities
+set_dropback_entity_status() {
+    timestamp=$(date '+%Y-%m-%dT%H:%M:%S')
+    curl --silent -X POST -H "Authorization: Bearer ${SUPERVISOR_TOKEN}" -H "Content-Type: application/json" http://supervisor/core/api/states/dropback.status -d '{"state":"'"$1"'", "attributes":{"friendly_name":"Dropback Status", "timestamp":"'"$timestamp"'"}}'
+}
+set_dropback_entity_sync() {
+    timestamp=$(date '+%Y-%m-%dT%H:%M:%S')
+    curl --silent -X POST -H "Authorization: Bearer ${SUPERVISOR_TOKEN}" -H "Content-Type: application/json" http://supervisor/core/api/states/dropback.sync -d '{"state":"'"$1"'", "attributes":{"friendly_name":"Dropback Sync", "timestamp":"'"$timestamp"'"}}'
+}
+
 # configure Dropbox access
 bashio::log.info "Initializing Dropback"
+set_dropback_entity_status "OK"
+set_dropback_entity_sync "None"
+
 if [ ! -e "$CONFIG_FILE" ]; then
     bashio::log.info "No config file found, requesting long lived Refresh Token..."
 
@@ -82,6 +95,7 @@ if [ ! -e "$CONFIG_FILE" ]; then
         bashio::log.fatal "Error getting Refresh Token"
         bashio::log.fatal "$ERROR $ERROR_DESCRIPTION"
         warn_about_access_token
+        set_dropback_entity_status "Error"
         bashio::exit.nok
     else
         # ensure app has correct permissions
@@ -92,6 +106,7 @@ if [ ! -e "$CONFIG_FILE" ]; then
             bashio::log.fatal "Missing scope \"$REQUIRED_SCOPE\""
             bashio::log.fatal "Please ensure the app has scope \"$REQUIRED_SCOPE\" enabled in the \"Permissions\" tab on dropbox.com/developers/apps"
             warn_about_access_token
+            set_dropback_entity_status "Error"
             bashio::exit.nok
         fi
     fi
@@ -120,6 +135,7 @@ else
     [ "$RESPONSE" != "" ] && bashio::log.fatal "$RESPONSE"
     warn_about_access_token
     rm "$CONFIG_FILE"
+    set_dropback_entity_status "Error"
     bashio::exit.nok
 fi
 
@@ -172,19 +188,23 @@ while read -r INPUT; do
             find $BACKUP_DIR -maxdepth 1 -type f -name "*.tar" -print0 |
             while IFS= read -r -d '' FILE; do
                 FILE_PATH=$(get_dropbox_file_path "$FILE")
+                set_dropback_entity_sync "${FILE_PATH/\//}"
                 bashio::log.debug "Sync $FILE to $FILE_PATH on Dropbox"
                 check_network_access
                 EXIT_CODE=0
                 RESPONSE=$(./dropbox_uploader.sh -q -s -f $CONFIG_FILE upload "$FILE" "$FILE_PATH") || EXIT_CODE=$?
                 if [ $EXIT_CODE -eq 0 ]; then
                     bashio::log.info "Synced $FILE to $FILE_PATH on Dropbox"
+                    set_dropback_entity_status "OK"
                 else
                     bashio::log.fatal "Failed to sync $FILE to $FILE_PATH on Dropbox"
                     [ "$RESPONSE" != "" ] && bashio::log.fatal "$RESPONSE"
+                    set_dropback_entity_status "Error"
                 fi
                 EXIT_CODE=0
             done
             bashio::log.info "Syncing done"
+            set_dropback_entity_sync "None"
 
         fi
     else